+++ backend/testing/src/main/kotlin/io/tolgee/testing/assert.kt
@@ -1,6 +1,7 @@
 package io.tolgee.testing
 
 import io.tolgee.testing.assertions.Assertions
+import org.assertj.core.api.AbstractBooleanAssert
 import org.assertj.core.api.AbstractDateAssert
 import org.assertj.core.api.AbstractLongAssert
 import org.assertj.core.api.IterableAssert
@@ -13,3 +14,6 @@ inline val <reified T> T.assert: ObjectAssert<T>
 inline val Date?.assert: AbstractDateAssert<*> get() = Assertions.assertThat(this)
 inline val <reified T> Iterable<T>.assert: IterableAssert<T> get() = Assertions.assertThat(this)
 inline val Long.assert: AbstractLongAssert<*> get() = Assertions.assertThat(this)
+
+inline val Boolean?.assert: AbstractBooleanAssert<*>
+  get() = Assertions.assertThat(this)